feat: 컨트롤러 id를 Long으로 수정

+++ secretboard/src/main/java/com/siat/secretboard/board/controller/BoardController.java
@@ -37,7 +37,7 @@ public class BoardController {
     }
 
     @PutMapping("/{id}")
-    public void updateBoard(@PathVariable("id") int id, @RequestBody BoardRequestDTO board) {
+    public void updateBoard(@PathVariable("id") Long id, @RequestBody BoardRequestDTO board) {
         board.setBoardIdx(id);
         boardService.updateBoard(board);
     }
